add 13 funct, +modif

+++ ft_striter.c
@@ -2,7 +2,7 @@
 
 void	ft_striter(char *s, void (*f)(char *))
 {
-	size_t i;
+	unsigned int	i;
 
 	i = 0;
 	while (s[i] != '\0')
